fix: records with data type mismatching goes to DLQ instead of failing the connector (#27)
* fix: records with data type mismatching go to DLQ instead of failing the connector

fixes #26
this is a first impl. it could be optimized further, but I assume bad data are rare
and it already does the job as it is.

there is a potential issue with `inflightSinkRecords` referencing all in-flight
data. in theory, this can +- double memory consumption. but we need the original
SinkRecord so we can send them to DLQ.